blog_project/requirements.txt
Django==3.0.7
djangorestframework-simplejwt
djangorestframework
django-auth
djangorestframework-jwt
pillow